Refactor cases array in tests for stat
Before this commit `simple_operation_cases`
were organized as map
(table indexed not with numbers),
in Lua iteration over map does not occur in
the order in which the elements were specified
in the map.
But simple operation cases should be executed
in the order in which they are specified,
because, for example, if `replace()` is performed
before `insert()`, an error will be received.
Therefore, `simple_operation_cases` has been
refactored as table indexed with numbers.
Select cases has been refactored for consistency.

Part of #193